Seperate download & write threads (#115)
* Simplify download code

* A few extra guards to prevent errors

* Split download worker up to reduce method size

* Inherit ctx, use shared rate limiter

* Be smarter about decimal chapters

* Remove empty series dirs after download

* Take chapter progress into account, catch panics in download, wait for workers to finish on cancel

* Fix content not being added to dashboard without refresh

* Share I/O workers between chapters

* Don't duplicate code in download worker, set state to cleanup when waiting on IO Workers

* Remove some uneeded fragmentation

* Skip these tests for now

* Linter

* [skip ci] Update dep

* Update go version

* Rewrite pasloe queue system

The system now uses a worker and nice channels, and a global context.
Now everything gets cancelled nicely.

The gracefull shutdown is not working yet, it's not sending the stop
request back to pasloe to remove the content just yet.

* Hook up yoitsu, delete files on shutdown

* Fix unit tests

* CLeanup code a bit

* Linter

+++ providers/pasloe/mangadex/manga.go
@@ -56,8 +56,8 @@ type manga struct {
 
 	coverFactory CoverFactory
 
-	lastFoundChapter int
-	lastFoundVolume  int
+	lastFoundChapter float64
+	lastFoundVolume  float64
 	foundLastVolume  bool
 	foundLastChapter bool
